adding block verification + fixing errors + adding OverallBalance

+++ src/main/java/com/dusk/externalAPIs/blockstream/models/TransactionStatus.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class TransactionStatus {
 
     private Boolean confirmed;
-    private Integer block_height;
+    private Long block_height;
     private String block_hash;
     private Long block_time;
